Verificação dos campos obrigatorios de Login, esqucir minha senha

+++ scripts/login.js
@@ -5,43 +5,59 @@ const API_AUTH = "pessoa/auth"
 const emailField = document.getElementById("email");
 const passwordField = document.getElementById("password");
 const submit = document.getElementById("submit");
+const pessoa = {}
 
 submit.addEventListener("click", (e) => {
     e.preventDefault();
     enviar();
 })
 
+function constroiPessoa(){
+    pessoa.email = emailField.value
+    pessoa.senha = passwordField.value
+}
+
+function verificaCamposObrigatorios(){
+    if(pessoa.email  && pessoa.senha){
+        return true
+    } else {
+        return false
+    }
+}
 function enviar(){
-    const pessoaLogin = {};
-    pessoaLogin.email = emailField.value
-    pessoaLogin.senha = passwordField.value
-    let status;
-    fetch(`${BASE_URL_SERVER}${API_AUTH}`,{
-        method: "POST",
-        headers: { "Content-Type":"application/json"},
-        body: JSON.stringify(pessoaLogin)
-    })
-    .then(res => {
-        status = res.status;
-        return res.json();
-    })
-    .then(({ token, email }) => {
-        if(status == 200){
-            localStorage.setItem("email", email)
-            localStorage.setItem("token", token)
-            direcionamento()
-        }else if(status == 401) {
-            localStorage.clear()
-            alert("Usuário ou senha inválido")
-            emailField.value = ""
-            passwordField.value = ""
-        }else{
-            localStorage.clear()
-            alert("Ocorreu um erro ao logar, tente novamente mais tarde")
-            emailField.value = ""
-            passwordField.value = ""  
-        }
-    })
+    constroiPessoa(); 
+    if( verificaCamposObrigatorios()) {
+        let status;
+        fetch(`${BASE_URL_SERVER}${API_AUTH}`,{
+            method: "POST",
+            headers: { "Content-Type":"application/json"},
+            body: JSON.stringify(pessoa)
+        })
+        .then(res => {
+            status = res.status;
+            return res.json();
+        })
+        .then(({ token, email }) => {
+            if(status == 200){
+                localStorage.setItem("email", email)
+                localStorage.setItem("token", token)
+                direcionamento()
+            }else if(status == 401) {
+                localStorage.clear()
+                alert("Usuário ou senha inválido")
+                emailField.value = ""
+                passwordField.value = ""
+            }else{
+                localStorage.clear()
+                alert("Ocorreu um erro ao logar, tente novamente mais tarde")
+                emailField.value = ""
+                passwordField.value = ""  
+            }
+        })
+    }else{
+        window.alert('Campos obrigatórios não preenchidos')
+    }
+   
 }
 
 function direcionamento(){
